avancement du projet

+++ TP2_Bieres_PLAYOUST_VAZQUEZ/src/tp2_bieres_playoust_vazquez/BouteilleBiere.java
@@ -10,7 +10,7 @@ package tp2_bieres_playoust_vazquez;
  */
 public class BouteilleBiere {
     String Nom;
-    float degreAlcool;
+    double degreAlcool;
     String Brasserie;
     boolean ouverte;
     
@@ -18,7 +18,9 @@ public class BouteilleBiere {
         System.out.println("Bouteille de " + Nom +" (" + degreAlcool + " degres) \nBrasserie : " + Brasserie ) ;
     }
     
+    
 }
+
